vSphere_with_Tanzu/wcp_cluster_health/requirements.txt
coloredlogs==15.0.1
Requests==2.32.3
slackclient==2.9.4
